Move printing logic into a separate package
